+++ cfdm/data/abstract/compressedarray.py
@@ -1,5 +1,6 @@
 import numpy as np
 
+from ..netcdfindexer import netcdf_indexer
 from .array import Array
 
 
@@ -189,10 +190,15 @@ class CompressedArray(Array):
             )
             u[u_indices] = subarray[...]
 
-        if indices is Ellipsis:
-            return u
-
-        return self.get_subspace(u, indices, copy=True)
+        u = netcdf_indexer(
+            u,
+            mask=False,
+            unpack=False,
+            always_masked_array=False,
+            orthogonal_indexing=True,
+            copy=False,
+        )
+        return u[indices]
 
     def _first_or_last_element(self, indices):
         """Return the first or last element of the compressed array.