+++ cfdm/docstring/docstring.py
@@ -401,6 +401,59 @@ _docstring_substitution_definitions = {
     "{{init cell_dimension: `int`}}": """cell_dimension: `int`
                 The position of the *data* dimension that indexes the
                 cells, either ``0`` or ``1``.""",
+    # init mask
+    "{{init mask: `bool`, optional}}": """mask: `bool`, optional
+                If True (the default) then mask by convention when
+                reading data from disk.
+
+                A netCDF array is masked depending on the values of
+                any of the netCDF attributes ``_FillValue``,
+                ``missing_value``, ``_Unsigned``, ``valid_min``,
+                ``valid_max``, and ``valid_range``.""",
+    # init unpack
+    "{{init unpack: `bool`, optional}}": """unpack: `bool`, optional
+                If True (the default) then unpack by convention when
+                reading data from disk.
+
+                A netCDF array is unpacked depending on the values of
+                the netCDF attributes ``add_offset`` and
+                ``scale_factor``.""",
+    # init attributes
+    "{{init attributes: `dict` or `None`, optional}}": """attributes: `dict` or `None`, optional
+                Provide netCDF attributes for the data as a dictionary
+                of key/value pairs.""",
+    # init storage_options
+    "{{init storage_options: `dict` or `None`, optional}}": """storage_options: `dict` or `None`, optional
+                Key/value pairs to be passed on to the creation of
+                `s3fs.S3FileSystem` file systems to control the
+                opening of files in S3 object stores. Ignored for
+                files not in an S3 object store, i.e. those whose
+                names do not start with ``s3:``.
+
+                By default, or if `None`, then *storage_options* is
+                taken as ``{}``.
+
+                If the ``'endpoint_url'`` key is not in
+                *storage_options* or is not in a dictionary defined by
+                the ``'client_kwargs`` key (which is always the case
+                when *storage_options* is `None`), then one will be
+                automatically inserted for accessing an S3 file. For
+                example, for a file name of
+                ``'s3://store/data/file.nc'``, an ``'endpoint_url'``
+                key with value ``'https://store'`` would be created.
+
+                *Parameter example:*
+                  For a file name of ``'s3://store/data/file.nc'``,
+                  the following are equivalent: ``None``, ``{}``, and
+                  ``{'endpoint_url': 'https://store'}``,
+                  ``{'client_kwargs': {'endpoint_url':
+                  'https://store'}}``
+
+                *Parameter example:*
+                  ``{'key': 'scaleway-api-key...', 'secret':
+                  'scaleway-secretkey...', 'endpoint_url':
+                  'https://s3.fr-par.scw.cloud', 'client_kwargs':
+                  {'region_name': 'fr-par'}}``""",
     # ----------------------------------------------------------------
     # Method description susbstitutions (4 levels of indentataion)
     # ----------------------------------------------------------------