+++ cfdm/read_write/netcdf/flatten/__init__.py
@@ -1,0 +1,16 @@
+"""Flatten NetCDF groups.
+
+Portions of this package were adapted from the `netcdf_flattener`
+library, which carries the following Apache 2.0 License:
+
+Copyright (c) 2020 EUMETSAT
+
+Licensed to the Apache Software Foundation (ASF) under one or more
+contributor license agreements. The ASF licenses this file to you
+under the Apache License, Version 2.0 (the "License"); you may not use
+this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+"""
+
+from .flatten import netcdf_flatten