+++ docs/source/conf.py
@@ -131,6 +131,7 @@ intersphinx_mapping = {
     "numpy": ("https://numpy.org/doc/stable", None),
     # 'netCDF4': ('https://unidata.github.io/netcdf4-python/', None),
     "cftime": ("https://unidata.github.io/cftime", None),
+    "h5netcdf": ("https://h5netcdf.org", None),
 }
 
 # This extension is meant to help with the common pattern of having